Conflicts:
	src/CommandingOfficers/Commander.java
	src/CommandingOfficers/CommanderAbility.java
	src/CommandingOfficers/CommanderLibrary.java
	src/UI/Art/SpriteArtist/CommanderOverlayArtist.java
	src/UI/Art/SpriteArtist/SpriteCOSetupArtist.java
	src/UI/Art/SpriteArtist/SpriteLibrary.java
	src/UI/COSetupController.java

+++ src/CommandingOfficers/CommanderAbility.java
@@ -11,9 +11,10 @@ public abstract class CommanderAbility
   protected String myName;
   private double myPowerCost;
   public final double baseCost;
+
   public int AIFlags = PHASE_TURN_START;
 
-  public CommanderAbility(Commander commander, String abilityName, int powerCost)
+  public CommanderAbility(Commander commander, String abilityName, double powerCost)
   {
     myCommander = commander;
     myName = abilityName;
